+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. Version number will be decided later -->
+- feat: Add `throttle.finish(true);` signature to cancel the trailing call
 
 ## 3.5.0